Compatibility clas for PGB
