+++ pyproject.toml
@@ -4,7 +4,7 @@
 [project]
 name = "postgresql-charms-single-kernel"
 description = "Shared and reusable code for PostgreSQL-related charms"
-version = "16.1.11"
+version = "16.1.12"
 readme = "README.md"
 license = {file = "LICENSE"}
 authors = [
@@ -64,7 +64,7 @@ log_cli_level = "INFO"
 [tool.ruff]
 # preview and explicit preview are enabled for CPY001
 preview = true
-target-version = "py38"
+target-version = "py312"
 src = ["src", "."]
 line-length = 99
 
@@ -111,6 +111,7 @@ convention = "google"
 
 [tool.ty.environment]
 python = ".tox/lint/"
+python-version = "3.12"
 
 [tool.ty.src]
 exclude = ["tests"]
