+++ single_kernel_postgresql/compat/__init__.py
@@ -1,0 +1,3 @@
+# Copyright 2025 Canonical Ltd.
+# See LICENSE file for licensing details.
+"""Utils and helpers for PostgreSQL charms."""
