+++ single_kernel_postgresql/utils/postgresql.py
@@ -22,21 +22,35 @@ Any charm using this library should import the `psycopg2` or `psycopg2-binary` d
 import logging
 import os
 import pwd
-from collections import OrderedDict
-from datetime import datetime, timezone
-from typing import Dict, List, Optional, Set, Tuple
+from datetime import UTC, datetime
 
 import psycopg2
-import psycopg2.errors
-import psycopg2.extensions
 from ops import ConfigData
 from psycopg2.sql import SQL, Identifier, Literal
 
+from ..compat.postgresql import (
+    ACCESS_GROUP_RELATION,
+    ALLOWED_ROLES,
+    ROLE_ADMIN,
+    ROLE_BACKUP,
+    ROLE_DATABASES_OWNER,
+    ROLE_DBA,
+    ROLE_DML,
+    ROLE_READ,
+    ROLE_STATS,
+    PostgreSQLBase,
+    PostgreSQLBaseError,
+    PostgreSQLCreateDatabaseError,  # noqa: F401
+    PostgreSQLCreateUserError,  # noqa: F401
+    PostgreSQLDeleteUserError,  # noqa: F401
+    PostgreSQLEnableDisableExtensionError,  # noqa: F401
+    PostgreSQLListUsersError,
+    PostgreSQLUndefinedHostError,  # noqa: F401
+    PostgreSQLUndefinedPasswordError,  # noqa: F401
+)
 from ..config.literals import (
-    BACKUP_USER,
     POSTGRESQL_STORAGE_PERMISSIONS,
     SNAP_USER,
-    SYSTEM_USERS,
     Substrates,
 )
 from .filesystem import change_owner, is_tmpfs
@@ -44,7 +58,6 @@ from .filesystem import change_owner, is_tmpfs
 # Groups to distinguish HBA access
 ACCESS_GROUP_IDENTITY = "identity_access"
 ACCESS_GROUP_INTERNAL = "internal_access"
-ACCESS_GROUP_RELATION = "relation_access"
 
 # List of access groups to filter role assignments by
 ACCESS_GROUPS = [
@@ -53,91 +66,23 @@ ACCESS_GROUPS = [
     ACCESS_GROUP_RELATION,
 ]
 
-ROLE_STATS = "charmed_stats"
-ROLE_READ = "charmed_read"
-ROLE_DML = "charmed_dml"
-ROLE_BACKUP = "charmed_backup"
-ROLE_DBA = "charmed_dba"
-ROLE_ADMIN = "charmed_admin"
-ROLE_DATABASES_OWNER = "charmed_databases_owner"
-ALLOWED_ROLES = {
-    ROLE_STATS,
-    ROLE_READ,
-    ROLE_DML,
-    ROLE_ADMIN,
-}
-
-INVALID_DATABASE_NAME_BLOCKING_MESSAGE = "invalid database name"
-INVALID_DATABASE_NAMES = ["databases", "postgres", "template0", "template1"]
-INVALID_EXTRA_USER_ROLE_BLOCKING_MESSAGE = "invalid role(s) for extra user roles"
-
-REQUIRED_PLUGINS = {
-    "address_standardizer": ["postgis"],
-    "address_standardizer_data_us": ["postgis"],
-    "jsonb_plperl": ["plperl"],
-    "postgis_raster": ["postgis"],
-    "postgis_tiger_geocoder": ["postgis", "fuzzystrmatch"],
-    "postgis_topology": ["postgis"],
-}
-DEPENDENCY_PLUGINS = set()
-for dependencies in REQUIRED_PLUGINS.values():
-    DEPENDENCY_PLUGINS |= set(dependencies)
-
 logger = logging.getLogger(__name__)
-
-
-class PostgreSQLBaseError(Exception):
-    """Base lib exception."""
-
-    message = None
 
 
 class PostgreSQLAssignGroupError(PostgreSQLBaseError):
     """Exception raised when assigning to a group fails."""
 
 
-class PostgreSQLCreateDatabaseError(PostgreSQLBaseError):
-    """Exception raised when creating a database fails."""
-
-    def __init__(self, message: Optional[str] = None):
-        super().__init__(message)
-        self.message = message
-
-
 class PostgreSQLCreateGroupError(PostgreSQLBaseError):
     """Exception raised when creating a group fails."""
-
-
-class PostgreSQLCreateUserError(PostgreSQLBaseError):
-    """Exception raised when creating a user fails."""
-
-    def __init__(self, message: Optional[str] = None):
-        super().__init__(message)
-        self.message = message
 
 
 class PostgreSQLUpdateUserError(PostgreSQLBaseError):
     """Exception raised when creating a user fails."""
 
 
-class PostgreSQLUndefinedHostError(PostgreSQLBaseError):
-    """Exception when host is not set."""
-
-
-class PostgreSQLUndefinedPasswordError(PostgreSQLBaseError):
-    """Exception when password is not set."""
-
-
 class PostgreSQLDatabasesSetupError(PostgreSQLBaseError):
     """Exception raised when the databases setup fails."""
-
-
-class PostgreSQLDeleteUserError(PostgreSQLBaseError):
-    """Exception raised when deleting a user fails."""
-
-
-class PostgreSQLEnableDisableExtensionError(PostgreSQLBaseError):
-    """Exception raised when enabling/disabling an extension fails."""
 
 
 class PostgreSQLGetLastArchivedWALError(PostgreSQLBaseError):
@@ -162,10 +107,6 @@ class PostgreSQLListAccessibleDatabasesForUserError(PostgreSQLBaseError):
 
 class PostgreSQLListGroupsError(PostgreSQLBaseError):
     """Exception raised when retrieving PostgreSQL groups list fails."""
-
-
-class PostgreSQLListUsersError(PostgreSQLBaseError):
-    """Exception raised when retrieving PostgreSQL users list fails."""
 
 
 class PostgreSQLUpdateUserPasswordError(PostgreSQLBaseError):
@@ -228,18 +169,18 @@ class PostgreSQLGrantDatabasePrivilegesToUserError(PostgreSQLBaseError):
     """Exception raised when granting database privileges to user."""
 
 
-class PostgreSQL:
+class PostgreSQL(PostgreSQLBase):
     """Class to encapsulate all operations related to interacting with PostgreSQL instance."""
 
     def __init__(
         self,
         substrate: Substrates,
-        primary_host: Optional[str],
-        current_host: Optional[str],
+        primary_host: str | None,
+        current_host: str | None,
         user: str,
-        password: Optional[str],
+        password: str | None,
         database: str,
-        system_users: Optional[List[str]] = None,
+        system_users: list[str] | None = None,
     ):
         """Create a PostgreSQL helper.
 
@@ -252,63 +193,9 @@ class PostgreSQL:
             database: default database name.
             system_users: list of system users.
         """
+        super().__init__(primary_host, current_host, user, password, database)
         self.substrate = substrate
-        self.primary_host = primary_host
-        self.current_host = current_host
-        self.user = user
-        self.password = password
-        self.database = database
         self.system_users = system_users if system_users else []
-
-    def _configure_pgaudit(self, enable: bool) -> None:
-        connection = None
-        try:
-            connection = self._connect_to_database()
-            connection.autocommit = True
-            with connection.cursor() as cursor:
-                cursor.execute("RESET ROLE;")
-                if enable:
-                    cursor.execute("ALTER SYSTEM SET pgaudit.log = 'ROLE,DDL,MISC,MISC_SET';")
-                    cursor.execute("ALTER SYSTEM SET pgaudit.log_client TO off;")
-                    cursor.execute("ALTER SYSTEM SET pgaudit.log_parameter TO off;")
-                else:
-                    cursor.execute("ALTER SYSTEM RESET pgaudit.log;")
-                    cursor.execute("ALTER SYSTEM RESET pgaudit.log_client;")
-                    cursor.execute("ALTER SYSTEM RESET pgaudit.log_parameter;")
-                cursor.execute("SELECT pg_reload_conf();")
-        finally:
-            if connection is not None:
-                connection.close()
-
-    def _connect_to_database(
-        self, database: Optional[str] = None, database_host: Optional[str] = None
-    ) -> psycopg2.extensions.connection:
-        """Creates a connection to the database.
-
-        Args:
-            database: database to connect to (defaults to the database
-                provided when the object for this class was created).
-            database_host: host to connect to instead of the primary host.
-
-        Returns:
-             psycopg2 connection object.
-        """
-        host = database_host if database_host is not None else self.primary_host
-        if not host:
-            raise PostgreSQLUndefinedHostError("Host not set")
-        if not self.password:
-            raise PostgreSQLUndefinedPasswordError("Password not set")
-
-        dbname = database if database else self.database
-        logger.debug(
-            f"New DB connection: dbname='{dbname}' user='{self.user}' host='{host}' connect_timeout=1"
-        )
-        connection = psycopg2.connect(
-            f"dbname='{dbname}' user='{self.user}' host='{host}'"
-            f"password='{self.password}' connect_timeout=1"
-        )
-        connection.autocommit = True
-        return connection
 
     def create_access_groups(self) -> None:
         """Create access groups to distinguish HBA authentication methods."""
@@ -333,233 +220,17 @@ class PostgreSQL:
             if connection is not None:
                 connection.close()
 
-    def create_database(
-        self,
-        database: str,
-        plugins: Optional[List[str]] = None,
-    ) -> None:
-        """Creates a new database and grant privileges to a user on it.
-
-        Args:
-            database: database to be created.
-            plugins: extensions to enable in the new database.
-        """
-        # The limit of 49 characters for the database name is due to the usernames that
-        # are created for each database, which have the prefix `charmed_` and a suffix
-        # like `_owner`, which summed to the database name must not exceed PostgreSQL
-        # maximum identifier length (63 characters, which is, the prefix, 8 characters,
-        # + database name, 49 characters maximum, + suffix, 6 characters).
-        if len(database) > 49:
-            logger.error(f"Invalid database name (it must not exceed 49 characters): {database}.")
-            raise PostgreSQLCreateDatabaseError(INVALID_DATABASE_NAME_BLOCKING_MESSAGE)
-        if database in INVALID_DATABASE_NAMES:
-            logger.error(f"Invalid database name: {database}.")
-            raise PostgreSQLCreateDatabaseError(INVALID_DATABASE_NAME_BLOCKING_MESSAGE)
-        plugins = plugins if plugins else []
-        try:
-            connection = self._connect_to_database()
-            cursor = connection.cursor()
-            cursor.execute(
-                SQL("SELECT datname FROM pg_database WHERE datname={};").format(Literal(database))
-            )
-            if cursor.fetchone() is None:
-                cursor.execute(SQL("SET ROLE {};").format(Identifier(ROLE_DATABASES_OWNER)))
-                cursor.execute(SQL("CREATE DATABASE {};").format(Identifier(database)))
-                cursor.execute(
-                    SQL("REVOKE ALL PRIVILEGES ON DATABASE {} FROM PUBLIC;").format(
-                        Identifier(database)
-                    )
-                )
-            with self._connect_to_database(database=database) as conn, conn.cursor() as curs:
-                curs.execute(SQL("SET ROLE {};").format(Identifier(ROLE_DATABASES_OWNER)))
-                curs.execute(SQL("SELECT set_up_predefined_catalog_roles();"))
-        except psycopg2.Error as e:
-            logger.error(f"Failed to create database: {e}")
-            raise PostgreSQLCreateDatabaseError() from e
-
-        # Enable preset extensions
-        if plugins:
-            self.enable_disable_extensions(dict.fromkeys(plugins, True), database)
-
-    def create_user(
-        self,
-        user: str,
-        password: Optional[str] = None,
-        admin: bool = False,
-        replication: bool = False,
-        extra_user_roles: Optional[List[str]] = None,
-        database: Optional[str] = None,
-        can_create_database: bool = False,
-    ) -> None:
-        """Creates a database user.
-
-        Args:
-            user: user to be created.
-            password: password to be assigned to the user.
-            admin: whether the user should have additional admin privileges.
-            replication: whether the user should have replication privileges.
-            extra_user_roles: additional privileges and/or roles to be assigned to the user.
-            database: optional database to allow the user to connect to.
-            can_create_database: whether the user should be able to create databases.
-        """
-        try:
-            roles, privileges = self._process_extra_user_roles(user, extra_user_roles)
-
-            with self._connect_to_database() as connection, connection.cursor() as cursor:
-                # Create or update the user.
-                cursor.execute(
-                    SQL("SELECT TRUE FROM pg_roles WHERE rolname={};").format(Literal(user))
-                )
-                if cursor.fetchone() is not None:
-                    user_definition = "ALTER ROLE {} "
-                    altering = True
-                else:
-                    user_definition = "CREATE ROLE {} "
-                    altering = False
-                user_definition += f"WITH LOGIN{' SUPERUSER' if admin else ''}{' REPLICATION' if replication else ''} ENCRYPTED PASSWORD '{password}'"
-                if not altering:
-                    user_definition, connect_statements = self._adjust_user_definition(
-                        user, roles, database, user_definition
-                    )
-                    if can_create_database:
-                        user_definition += " CREATEDB"
-                    if privileges:
-                        user_definition += f" {' '.join(privileges)}"
-                else:
-                    db_roles, connect_statements = self._adjust_user_roles(user, roles, database)
-                    roles = [*db_roles, *roles]
-                cursor.execute(SQL("RESET ROLE;"))
-                cursor.execute(SQL("BEGIN;"))
-                cursor.execute(SQL("SET LOCAL log_statement = 'none';"))
-                cursor.execute(SQL(f"{user_definition};").format(Identifier(user)))
-                cursor.execute(SQL("COMMIT;"))
-                if len(connect_statements) > 0:
-                    for connect_statement in connect_statements:
-                        cursor.execute(connect_statement)
-
-                # Add extra user roles to the new user.
-                for role in roles:
-                    cursor.execute(
-                        SQL("GRANT {} TO {};").format(Identifier(role), Identifier(user))
-                    )
-        except psycopg2.Error as e:
-            logger.error(f"Failed to create user: {e}")
-            raise PostgreSQLCreateUserError() from e
-
-    def _adjust_user_definition(
-        self, user: str, roles: Optional[List[str]], database: Optional[str], user_definition: str
-    ) -> Tuple[str, List[str]]:
-        """Adjusts the user definition to include additional statements.
-
-        Returns:
-            A tuple containing the adjusted user definition and a list of additional statements.
-        """
-        db_roles, connect_statements = self._adjust_user_roles(user, roles, database)
-        if db_roles:
-            str_roles = [f'"{role}"' for role in db_roles]
-            user_definition += f" IN ROLE {', '.join(str_roles)}"
-        return user_definition, connect_statements
-
-    def _adjust_user_roles(
-        self, user: str, roles: Optional[List[str]], database: Optional[str]
-    ) -> Tuple[List[str], List[str]]:
-        """Adjusts the user definition to include additional statements.
-
-        Returns:
-            A tuple containing the adjusted user definition and a list of additional statements.
-        """
-        db_roles = []
-        connect_statements = []
-        if database:
-            if roles is not None and not any(
-                role in [ROLE_STATS, ROLE_READ, ROLE_DML, ROLE_BACKUP, ROLE_DBA] for role in roles
-            ):
-                db_roles.append(f"charmed_{database}_admin")
-                db_roles.append(f"charmed_{database}_dml")
-            else:
-                connect_statements.append(
-                    SQL("GRANT CONNECT ON DATABASE {} TO {};").format(
-                        Identifier(database), Identifier(user)
-                    )
-                )
-        if roles is not None and any(
-            role
-            in [
-                ROLE_STATS,
-                ROLE_READ,
-                ROLE_DML,
-                ROLE_BACKUP,
-                ROLE_DBA,
-                ROLE_ADMIN,
-                ROLE_DATABASES_OWNER,
-            ]
-            for role in roles
-        ):
-            for system_database in ["postgres", "template1"]:
-                connect_statements.append(
-                    SQL("GRANT CONNECT ON DATABASE {} TO {};").format(
-                        Identifier(system_database), Identifier(user)
-                    )
-                )
-        return db_roles, connect_statements
-
-    def _process_extra_user_roles(
-        self, user: str, extra_user_roles: Optional[List[str]] = None
-    ) -> Tuple[List[str], Set[str]]:
-        # Separate roles and privileges from the provided extra user roles.
-        roles = []
-        privileges = set()
-        if extra_user_roles:
-            if len(extra_user_roles) > 2 and sorted(extra_user_roles) != [
-                ROLE_ADMIN,
-                "createdb",
-                ACCESS_GROUP_RELATION,
-            ]:
-                extra_user_roles.remove(ACCESS_GROUP_RELATION)
-                logger.error(
-                    "Invalid extra user roles: "
-                    f"{', '.join(extra_user_roles)}. "
-                    f"Only 'createdb' and '{ROLE_ADMIN}' are allowed together."
-                )
-                raise PostgreSQLCreateUserError(INVALID_EXTRA_USER_ROLE_BLOCKING_MESSAGE)
-            valid_privileges, valid_roles = self.list_valid_privileges_and_roles()
-            roles = [
-                role
-                for role in extra_user_roles
-                if (
-                    user == BACKUP_USER
-                    or user in SYSTEM_USERS
-                    or role in valid_roles
-                    or role == ACCESS_GROUP_RELATION
-                    or role == "createdb"
-                )
-            ]
-            if "createdb" in extra_user_roles:
-                extra_user_roles.remove("createdb")
-                roles.remove("createdb")
-                extra_user_roles.append(ROLE_DATABASES_OWNER)
-                roles.append(ROLE_DATABASES_OWNER)
-            privileges = {
-                extra_user_role
-                for extra_user_role in extra_user_roles
-                if extra_user_role and extra_user_role not in roles
-            }
-            invalid_privileges = [
-                privilege for privilege in privileges if privilege not in valid_privileges
-            ]
-            if len(invalid_privileges) > 0:
-                logger.error(f"Invalid extra user roles: {', '.join(privileges)}")
-                raise PostgreSQLCreateUserError(INVALID_EXTRA_USER_ROLE_BLOCKING_MESSAGE)
-        return roles, privileges
-
     def create_predefined_instance_roles(self) -> None:
         """Create predefined instance roles."""
         connection = None
         try:
             for database in self._get_existing_databases():
-                with self._connect_to_database(
-                    database=database,
-                ) as connection, connection.cursor() as cursor:
+                with (
+                    self._connect_to_database(
+                        database=database,
+                    ) as connection,
+                    connection.cursor() as cursor,
+                ):
                     cursor.execute(SQL("CREATE EXTENSION IF NOT EXISTS set_user;"))
         finally:
             if connection is not None:
@@ -594,9 +265,12 @@ class PostgreSQL:
 
         try:
             for database in ["postgres", "template1"]:
-                with self._connect_to_database(
-                    database=database,
-                ) as connection, connection.cursor() as cursor:
+                with (
+                    self._connect_to_database(
+                        database=database,
+                    ) as connection,
+                    connection.cursor() as cursor,
+                ):
                     existing_roles = self.list_existing_roles()
                     for role, queries in role_to_queries.items():
                         for index, query in enumerate(queries):
@@ -615,7 +289,7 @@ class PostgreSQL:
                 connection.close()
 
     def grant_database_privileges_to_user(
-        self, user: str, database: str, privileges: List[str]
+        self, user: str, database: str, privileges: list[str]
     ) -> None:
         """Grant the specified privileges on the provided database for the user."""
         try:
@@ -628,45 +302,6 @@ class PostgreSQL:
         except psycopg2.Error as e:
             logger.error(f"Failed to grant privileges to user: {e}")
             raise PostgreSQLGrantDatabasePrivilegesToUserError() from e
-
-    def delete_user(self, user: str) -> None:
-        """Deletes a database user.
-
-        Args:
-            user: user to be deleted.
-        """
-        # First of all, check whether the user exists. Otherwise, do nothing.
-        users = self.list_users()
-        if user not in users:
-            return
-
-        # List all databases.
-        try:
-            with self._connect_to_database() as connection, connection.cursor() as cursor:
-                cursor.execute("SELECT datname FROM pg_database WHERE datistemplate = false;")
-                databases = [row[0] for row in cursor.fetchall()]
-
-            # Existing objects need to be reassigned in each database
-            # before the user can be deleted.
-            for database in databases:
-                with self._connect_to_database(
-                    database
-                ) as connection, connection.cursor() as cursor:
-                    cursor.execute(SQL("RESET ROLE;"))
-                    cursor.execute(
-                        SQL("REASSIGN OWNED BY {} TO {};").format(
-                            Identifier(user), Identifier(self.user)
-                        )
-                    )
-                    cursor.execute(SQL("DROP OWNED BY {};").format(Identifier(user)))
-
-            # Delete the user.
-            with self._connect_to_database() as connection, connection.cursor() as cursor:
-                cursor.execute(SQL("RESET ROLE;"))
-                cursor.execute(SQL("DROP ROLE {};").format(Identifier(user)))
-        except psycopg2.Error as e:
-            logger.error(f"Failed to delete user: {e}")
-            raise PostgreSQLDeleteUserError() from e
 
     def grant_internal_access_group_memberships(self) -> None:
         """Grant membership to the internal access-group to existing internal users."""
@@ -716,8 +351,8 @@ class PostgreSQL:
         self,
         user: str,
         database: str,
-        schematables: List[str],
-        old_schematables: Optional[List[str]] = None,
+        schematables: list[str],
+        old_schematables: list[str] | None = None,
     ) -> None:
         """Grant CONNECT privilege on database and SELECT privilege on tables.
 
@@ -760,7 +395,7 @@ class PostgreSQL:
                 connection.close()
 
     def revoke_replication_privileges(
-        self, user: str, database: str, schematables: List[str]
+        self, user: str, database: str, schematables: list[str]
     ) -> None:
         """Revoke all privileges from tables and database.
 
@@ -791,58 +426,6 @@ class PostgreSQL:
             if connection:
                 connection.close()
 
-    def enable_disable_extensions(
-        self, extensions: Dict[str, bool], database: Optional[str] = None
-    ) -> None:
-        """Enables or disables a PostgreSQL extension.
-
-        Args:
-            extensions: the name of the extensions.
-            database: optional database where to enable/disable the extension.
-
-        Raises:
-            PostgreSQLEnableDisableExtensionError if the operation fails.
-        """
-        connection = None
-        try:
-            if database is not None:
-                databases = [database]
-            else:
-                # Retrieve all the databases.
-                with self._connect_to_database() as connection, connection.cursor() as cursor:
-                    cursor.execute("SELECT datname FROM pg_database WHERE NOT datistemplate;")
-                    databases = {database[0] for database in cursor.fetchall()}
-
-            ordered_extensions = OrderedDict()
-            for plugin in DEPENDENCY_PLUGINS:
-                ordered_extensions[plugin] = extensions.get(plugin, False)
-            for extension, enable in extensions.items():
-                ordered_extensions[extension] = enable
-
-            self._configure_pgaudit(False)
-
-            # Enable/disabled the extension in each database.
-            for database in databases:
-                with self._connect_to_database(
-                    database=database
-                ) as connection, connection.cursor() as cursor:
-                    for extension, enable in ordered_extensions.items():
-                        cursor.execute(
-                            f"CREATE EXTENSION IF NOT EXISTS {extension};"
-                            if enable
-                            else f"DROP EXTENSION IF EXISTS {extension};"
-                        )
-            self._configure_pgaudit(ordered_extensions.get("pgaudit", False))
-        except psycopg2.errors.UniqueViolation:  # type: ignore
-            pass
-        except psycopg2.errors.DependentObjectsStillExist:  # type: ignore
-            raise
-        except psycopg2.Error as e:
-            raise PostgreSQLEnableDisableExtensionError() from e
-        finally:
-            if connection is not None:
-                connection.close()
-
     def get_last_archived_wal(self) -> str:
         """Get the name of the last archived wal for the current PostgreSQL cluster."""
         try:
@@ -865,41 +448,44 @@ class PostgreSQL:
             logger.error(f"Failed to get PostgreSQL current timeline id: {e}")
             raise PostgreSQLGetCurrentTimelineError() from e
 
-    def get_postgresql_text_search_configs(self) -> Set[str]:
+    def get_postgresql_text_search_configs(self) -> set[str]:
         """Returns the PostgreSQL available text search configs.
 
         Returns:
             Set of PostgreSQL text search configs.
         """
-        with self._connect_to_database(
-            database_host=self.current_host
-        ) as connection, connection.cursor() as cursor:
+        with (
+            self._connect_to_database(database_host=self.current_host) as connection,
+            connection.cursor() as cursor,
+        ):
             cursor.execute("SELECT CONCAT('pg_catalog.', cfgname) FROM pg_ts_config;")
             text_search_configs = cursor.fetchall()
             return {text_search_config[0] for text_search_config in text_search_configs}
 
-    def get_postgresql_timezones(self) -> Set[str]:
+    def get_postgresql_timezones(self) -> set[str]:
         """Returns the PostgreSQL available timezones.
 
         Returns:
             Set of PostgreSQL timezones.
         """
-        with self._connect_to_database(
-            database_host=self.current_host
-        ) as connection, connection.cursor() as cursor:
+        with (
+            self._connect_to_database(database_host=self.current_host) as connection,
+            connection.cursor() as cursor,
+        ):
             cursor.execute("SELECT name FROM pg_timezone_names;")
             timezones = cursor.fetchall()
             return {timezone[0] for timezone in timezones}
 
-    def get_postgresql_default_table_access_methods(self) -> Set[str]:
+    def get_postgresql_default_table_access_methods(self) -> set[str]:
         """Returns the PostgreSQL available table access methods.
 
         Returns:
             Set of PostgreSQL table access methods.
         """
-        with self._connect_to_database(
-            database_host=self.current_host
-        ) as connection, connection.cursor() as cursor:
+        with (
+            self._connect_to_database(database_host=self.current_host) as connection,
+            connection.cursor() as cursor,
+        ):
             cursor.execute("SELECT amname FROM pg_am WHERE amtype = 't';")
             access_methods = cursor.fetchall()
             return {access_method[0] for access_method in access_methods}
@@ -912,9 +498,10 @@ class PostgreSQL:
         """
         host = self.current_host if current_host else None
         try:
-            with self._connect_to_database(
-                database_host=host
-            ) as connection, connection.cursor() as cursor:
+            with (
+                self._connect_to_database(database_host=host) as connection,
+                connection.cursor() as cursor,
+            ):
                 cursor.execute("SELECT version();")
                 # Split to get only the version number. There should always be a version.
                 return cursor.fetchone()[0].split(" ")[1]
@@ -933,9 +520,12 @@ class PostgreSQL:
             whether TLS is enabled.
         """
         try:
-            with self._connect_to_database(
-                database_host=self.current_host if check_current_host else None
-            ) as connection, connection.cursor() as cursor:
+            with (
+                self._connect_to_database(
+                    database_host=self.current_host if check_current_host else None
+                ) as connection,
+                connection.cursor() as cursor,
+            ):
                 cursor.execute("SHOW ssl;")
                 # SSL state should always be set
                 return "on" in cursor.fetchone()[0]
@@ -943,7 +533,7 @@ class PostgreSQL:
             # Connection errors happen when PostgreSQL has not started yet.
             return False
 
-    def list_access_groups(self, current_host=False) -> Set[str]:
+    def list_access_groups(self, current_host=False) -> set[str]:
         """Returns the list of PostgreSQL database access groups.
 
         Args:
@@ -956,9 +546,10 @@ class PostgreSQL:
         connection = None
         host = self.current_host if current_host else None
         try:
-            with self._connect_to_database(
-                database_host=host
-            ) as connection, connection.cursor() as cursor:
+            with (
+                self._connect_to_database(database_host=host) as connection,
+                connection.cursor() as cursor,
+            ):
                 cursor.execute(
                     "SELECT groname FROM pg_catalog.pg_group WHERE groname LIKE '%_access';"
                 )
@@ -971,7 +562,7 @@ class PostgreSQL:
             if connection is not None:
                 connection.close()
 
-    def list_accessible_databases_for_user(self, user: str, current_host=False) -> Set[str]:
+    def list_accessible_databases_for_user(self, user: str, current_host=False) -> set[str]:
         """Returns the list of accessible databases for a specific user.
 
         Args:
@@ -986,9 +577,10 @@ class PostgreSQL:
         connection = None
         host = self.current_host if current_host else None
         try:
-            with self._connect_to_database(
-                database_host=host
-            ) as connection, connection.cursor() as cursor:
+            with (
+                self._connect_to_database(database_host=host) as connection,
+                connection.cursor() as cursor,
+            ):
                 cursor.execute(
                     SQL(
                         "SELECT TRUE FROM pg_catalog.pg_user WHERE usename = {} AND usesuper;"
@@ -1010,7 +602,7 @@ class PostgreSQL:
             if connection is not None:
                 connection.close()
 
-    def list_users(self, group: Optional[str] = None, current_host=False) -> Set[str]:
+    def list_users(self, group: str | None = None, current_host=False) -> set[str]:
         """Returns the list of PostgreSQL database users.
 
         Args:
@@ -1024,9 +616,10 @@ class PostgreSQL:
         connection = None
         host = self.current_host if current_host else None
         try:
-            with self._connect_to_database(
-                database_host=host
-            ) as connection, connection.cursor() as cursor:
+            with (
+                self._connect_to_database(database_host=host) as connection,
+                connection.cursor() as cursor,
+            ):
                 if group:
                     query = SQL(
                         "SELECT usename FROM (SELECT UNNEST(grolist) AS user_id FROM pg_catalog.pg_group WHERE groname = {}) AS g JOIN pg_catalog.pg_user AS u ON g.user_id = u.usesysid;"
@@ -1043,7 +636,7 @@ class PostgreSQL:
             if connection is not None:
                 connection.close()
 
-    def list_users_from_relation(self, current_host=False) -> Set[str]:
+    def list_users_from_relation(self, current_host=False) -> set[str]:
         """Returns the list of PostgreSQL database users that were created by a relation.
 
         Args:
@@ -1056,9 +649,10 @@ class PostgreSQL:
         connection = None
         host = self.current_host if current_host else None
         try:
-            with self._connect_to_database(
-                database_host=host
-            ) as connection, connection.cursor() as cursor:
+            with (
+                self._connect_to_database(database_host=host) as connection,
+                connection.cursor() as cursor,
+            ):
                 cursor.execute(
                     "SELECT usename "
                     "FROM pg_catalog.pg_user "
@@ -1075,7 +669,7 @@ class PostgreSQL:
             if connection is not None:
                 connection.close()
 
-    def list_existing_roles(self) -> Set[str]:
+    def list_existing_roles(self) -> set[str]:
         """Returns a set containing the existing roles.
 
         Returns:
@@ -1085,7 +679,7 @@ class PostgreSQL:
             cursor.execute("SELECT rolname FROM pg_roles;")
             return {role[0] for role in cursor.fetchall() if role[0]}
 
-    def list_valid_privileges_and_roles(self) -> Tuple[Set[str], Set[str]]:
+    def list_valid_privileges_and_roles(self) -> tuple[set[str], set[str]]:
         """Returns two sets with valid privileges and roles.
 
         Returns:
@@ -1096,7 +690,7 @@ class PostgreSQL:
             "superuser",
         }, ALLOWED_ROLES
 
-    def _get_existing_databases(self) -> List[str]:
+    def _get_existing_databases(self) -> list[str]:
         # Template1 should go first
         databases = ["template1"]
         connection = None
@@ -1136,7 +730,7 @@ class PostgreSQL:
             # Rename existing temp tablespace instead of dropping it.
             # Timestamp collision is not possible: the charm ensures this code runs leader-only,
             # and it executes within a single database transaction holding exclusive locks.
-            new_name = f"temp_{datetime.now(timezone.utc).strftime('%Y%m%d%H%M%S')}"
+            new_name = f"temp_{datetime.now(UTC).strftime('%Y%m%d%H%M%S')}"
             cursor.execute(f"ALTER TABLESPACE temp RENAME TO {new_name};")
 
             # List temp tablespaces with suffix for operator follow-up cleanup
@@ -1156,7 +750,7 @@ class PostgreSQL:
                 temp_location,
             )
 
-    def set_up_database(self, temp_location: Optional[str] = None) -> None:
+    def set_up_database(self, temp_location: str | None = None) -> None:
         """Set up postgres database with the right permissions.
 
         This method configures the postgres database with appropriate permissions and
@@ -1211,9 +805,10 @@ class PostgreSQL:
             connection.close()
             connection = None
 
-            with self._connect_to_database(
-                database="template1"
-            ) as connection, connection.cursor() as cursor:
+            with (
+                self._connect_to_database(database="template1") as connection,
+                connection.cursor() as cursor,
+            ):
                 cursor.execute(
                     f"SELECT TRUE FROM pg_roles WHERE rolname='{ROLE_DATABASES_OWNER}';"  # noqa: S608
                 )
@@ -1297,9 +892,10 @@ $$ LANGUAGE plpgsql;"""  # noqa: S608
         connection = None
         try:
             for database in self._get_existing_databases():
-                with self._connect_to_database(
-                    database=database
-                ) as connection, connection.cursor() as cursor:
+                with (
+                    self._connect_to_database(database=database) as connection,
+                    connection.cursor() as cursor,
+                ):
                     cursor.execute(SQL("CREATE EXTENSION IF NOT EXISTS login_hook;"))
                     cursor.execute(SQL("CREATE SCHEMA IF NOT EXISTS login_hook;"))
                     cursor.execute(SQL(function_creation_statement))
@@ -1405,9 +1001,10 @@ $$ LANGUAGE plpgsql security definer;"""  # noqa: S608
         connection = None
         try:
             for database in self._get_existing_databases():
-                with self._connect_to_database(
-                    database=database
-                ) as connection, connection.cursor() as cursor:
+                with (
+                    self._connect_to_database(database=database) as connection,
+                    connection.cursor() as cursor,
+                ):
                     cursor.execute(SQL(function_creation_statement))
                     cursor.execute(
                         SQL("ALTER FUNCTION set_up_predefined_catalog_roles OWNER TO operator;")
@@ -1435,7 +1032,7 @@ $$ LANGUAGE plpgsql security definer;"""  # noqa: S608
                 connection.close()
 
     def update_user_password(
-        self, username: str, password: str, database_host: Optional[str] = None
+        self, username: str, password: str, database_host: str | None = None
     ) -> None:
         """Update a user password.
 
@@ -1449,9 +1046,10 @@ $$ LANGUAGE plpgsql security definer;"""  # noqa: S608
         """
         connection = None
         try:
-            with self._connect_to_database(
-                database_host=database_host
-            ) as connection, connection.cursor() as cursor:
+            with (
+                self._connect_to_database(database_host=database_host) as connection,
+                connection.cursor() as cursor,
+            ):
                 cursor.execute(SQL("RESET ROLE;"))
                 cursor.execute(SQL("BEGIN;"))
                 cursor.execute(SQL("SET LOCAL log_statement = 'none';"))
@@ -1521,7 +1119,7 @@ $$ LANGUAGE plpgsql security definer;"""  # noqa: S608
             if connection:
                 connection.close()
 
-    def create_publication(self, db: str, name: str, schematables: List[str]) -> None:
+    def create_publication(self, db: str, name: str, schematables: list[str]) -> None:
         """Create PostgreSQL publication."""
         connection = None
         try:
@@ -1562,7 +1160,7 @@ $$ LANGUAGE plpgsql security definer;"""  # noqa: S608
             if connection:
                 connection.close()
 
-    def alter_publication(self, db: str, name: str, schematables: List[str]) -> None:
+    def alter_publication(self, db: str, name: str, schematables: list[str]) -> None:
         """Alter PostgreSQL publication."""
         connection = None
         try:
@@ -1719,7 +1317,7 @@ $$ LANGUAGE plpgsql security definer;"""  # noqa: S608
                 connection.close()
 
     @staticmethod
-    def build_postgresql_group_map(group_map: Optional[str]) -> List[Tuple]:
+    def build_postgresql_group_map(group_map: str | None) -> list[tuple]:
         """Build the PostgreSQL authorization group-map.
 
         Args:
@@ -1756,8 +1354,8 @@ $$ LANGUAGE plpgsql security definer;"""  # noqa: S608
 
     @staticmethod
     def build_postgresql_parameters(
-        config_options: ConfigData, available_memory: int, limit_memory: Optional[int] = None
-    ) -> Optional[dict]:
+        config_options: ConfigData, available_memory: int, limit_memory: int | None = None
+    ) -> dict | None:
         """Builds the PostgreSQL parameters.
 
         Args:
@@ -1827,9 +1425,10 @@ $$ LANGUAGE plpgsql security definer;"""  # noqa: S608
             Whether the date style is valid.
         """
         try:
-            with self._connect_to_database(
-                database_host=self.current_host
-            ) as connection, connection.cursor() as cursor:
+            with (
+                self._connect_to_database(database_host=self.current_host) as connection,
+                connection.cursor() as cursor,
+            ):
                 cursor.execute(
                     SQL(
                         "SET DateStyle to {};",
@@ -1839,7 +1438,7 @@ $$ LANGUAGE plpgsql security definer;"""  # noqa: S608
         except psycopg2.Error:
             return False
 
-    def validate_group_map(self, group_map: Optional[str]) -> bool:
+    def validate_group_map(self, group_map: str | None) -> bool:
         """Validate the PostgreSQL authorization group-map.
 
         Args:
@@ -1870,26 +1469,6 @@ $$ LANGUAGE plpgsql security definer;"""  # noqa: S608
 
         return True
 
-    def is_user_in_hba(self, username: str) -> bool:
-        """Check if user was added in pg_hba."""
-        connection = None
-        try:
-            with self._connect_to_database() as connection, connection.cursor() as cursor:
-                cursor.execute(
-                    SQL(
-                        "SELECT COUNT(*) FROM pg_hba_file_rules WHERE {} = ANY(user_name);"
-                    ).format(Literal(username))
-                )
-                if result := cursor.fetchone():
-                    return result[0] > 0
-                return False
-        except psycopg2.Error as e:
-            logger.debug(f"Failed to check pg_hba: {e}")
-            return False
-        finally:
-            if connection:
-                connection.close()
-
     def drop_hba_triggers(self) -> None:
         """Drop pg_hba triggers on schema change."""
         try:
@@ -1912,9 +1491,10 @@ $$ LANGUAGE plpgsql security definer;"""  # noqa: S608
 
         for database in databases:
             try:
-                with self._connect_to_database(
-                    database
-                ) as connection, connection.cursor() as cursor:
+                with (
+                    self._connect_to_database(database) as connection,
+                    connection.cursor() as cursor,
+                ):
                     cursor.execute(
                         SQL("DROP EVENT TRIGGER IF EXISTS update_pg_hba_on_create_schema;")
                     )
@@ -1927,7 +1507,7 @@ $$ LANGUAGE plpgsql security definer;"""  # noqa: S608
                 if connection:
                     connection.close()
 
-    def list_databases(self, prefix: Optional[str] = None) -> List[str]:
+    def list_databases(self, prefix: str | None = None) -> list[str]:
         """List non-system databases starting with prefix."""
         prefix_stmt = (
             SQL(" AND datname LIKE {}").format(Literal(prefix + "%")) if prefix else SQL("")
@@ -1947,7 +1527,7 @@ $$ LANGUAGE plpgsql security definer;"""  # noqa: S608
                 connection.close()
 
     def add_user_to_databases(
-        self, user: str, databases: List[str], extra_user_roles: Optional[List[str]] = None
+        self, user: str, databases: list[str], extra_user_roles: list[str] | None = None
     ) -> None:
         """Grant user access to a database."""
         try:
@@ -1974,7 +1554,7 @@ $$ LANGUAGE plpgsql security definer;"""  # noqa: S608
             logger.error(f"Failed to add user: {e}")
             raise PostgreSQLUpdateUserError() from e
 
-    def remove_user_from_databases(self, user: str, databases: List[str]) -> None:
+    def remove_user_from_databases(self, user: str, databases: list[str]) -> None:
         """Revoke user access to a database."""
         try:
             for database in databases:
